Comment out Undeployed Server Validation Tests
In the VMConfigurationData.psd1 file, the NanoServer and PullServer are commented out and not configured. As these servers are not deployed, all of the tests to validate these server configurations fail during the Validate-Lab phase causing the system to enter an infinite loop. Either the servers need to be deployed, or the tests need to be removed.